ai-bias-search/pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-bias-search"
version = "0.1.0"
description = "Production-ready scaffold for comparing literature search platforms."
authors = ["AI Bias Search Team <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ai_bias_search" }]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
pydantic = "^2.6.0"
pandas = "^2.1.0"
pyarrow = "^14.0.0"
tenacity = "^8.2.3"
python-dotenv = "^1.0.0"
matplotlib = "^3.8.0"
jinja2 = "^3.1.3"
scipy = "^1.11.0"
numpy = "^1.26.0"
typer = { version = "^0.9.0", extras = ["all"] }
diskcache = "^5.6.3"
upsetplot = "^0.9.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
respx = "^0.20.2"
pytest-mock = "^3.12.0"
ruff = "^0.3.0"
black = "^24.3.0"
mypy = "^1.6.0"

[tool.poetry.scripts]
ai-bias-search = "ai_bias_search.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
